requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.1
django-crispy-forms==1.9.2
django-filter==2.4.0
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
openpyxl==2.6.0
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
xlwt==1.3.0
